Update diff-quality to be compatible with Pylint 1.0
The new Pylint changes to allow a --message-format option, which makes
parsing a bit easier. Unfortunately it doesn't seem like we can get
the numeric ID of the error, but that's not a huge loss. This change
keeps the report format as close as possible to the existing one, and
updates tests to handle the lack of IDs.